Updated rules, removed panther_azure_helpers and added to panther_msft_helpers
