--- global_helpers/panther_azure_helpers.yml
@@ -1,5 +1,0 @@
-AnalysisType: global
-Filename: panther_azure_helpers.py
-GlobalID: "panther_azure_helpers"
-Description: >
-  Global helpers for Azure detections
