+++ rules/azure_entraid_rules/azure_role_changed_pim.py
@@ -1,10 +1,9 @@
-from panther_azure_helpers import azure_rule_context, azure_success, get_target_name
-from panther_base_helpers import deep_walk
+from panther_msft_helpers import azure_rule_context, azure_success, get_target_name
 
 
 def rule(event):
     operation = event.get("operationName", default="")
-    if azure_success and "Add member to role in PIM completed" in operation:
+    if azure_success(event) and "Add member to role in PIM completed" in operation:
         return True
 
     return False
@@ -16,8 +15,8 @@ def title(event):
         "properties", "initiatedBy", "user", "userPrincipalName", default=""
     )
     target_name = get_target_name(event)
-    role = deep_walk(
-        event, "properties", "targetResources", "displayName", return_val="first", default=""
+    role = event.deep_walk(
+        "properties", "targetResources", "displayName", return_val="first", default=""
     )
     return f"{actor_name} added {target_name} as {role} successfully with {operation_name}"
